requirements.txt
flask
flask-mail
flask-cors
python-dotenv
gunicorn
